+++ src/firebase.js
@@ -1,0 +1,22 @@
+// src/firebase.js
+import { initializeApp } from "firebase/app";
+import { getAuth } from "firebase/auth";
+import { getFirestore } from "firebase/firestore";
+
+// Config-ul tău de la Firebase Console
+const firebaseConfig = {
+  apiKey: "API_KEY",
+  authDomain: "PROJECT_ID.firebaseapp.com",
+  projectId: "PROJECT_ID",
+  storageBucket: "PROJECT_ID.appspot.com",
+  messagingSenderId: "SENDER_ID",
+  appId: "APP_ID",
+};
+
+// Initialize Firebase
+const app = initializeApp(firebaseConfig);
+
+// Servicii Firebase pe care le vei folosi
+export const auth = getAuth(app);
+export const db = getFirestore(app);
+export default app;
